Move test to GroupTests, refactor
